orchael-sdk/pyproject.toml
[project]
name = "orchael-sdk"
version = "0.1.0"
description = "A framework for building chat processors and chat applications"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Orchael Team"}
]
license = {text = "MIT"}
keywords = ["chat", "processor", "sdk", "framework"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Communications :: Chat",
]
dependencies = [
    "typing-extensions>=4.0.0",
    "click>=8.0.0",
    "PyYAML>=6.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0"
]

[project.scripts]
orchael-sdk-cli = "orchael_sdk.cli:main"
orchael-sdk-server = "orchael_sdk.cli:server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["orchael_sdk"]

[tool.hatch.build.targets.sdist]
include = [
    "/orchael_sdk",
    "/README.md",
    "/pyproject.toml"
]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]

[tool.ruff.format]
indent-style = "space"

[tool.mypy]
strict = true
warn_return_any = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "--cov=orchael_sdk",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests"
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning"
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.10",
    "types-pyyaml>=6.0.12.20250822",
    "httpx>=0.25.0",
    "types-requests>=2.31.0",
    "types-requests>=2.31.0",
    "pydantic>=2.0.0",
]
